[dev-2991] moving logic for building AwardAmountsTable to the component itself
